Cloud Dev/Final Proj/requirements.txt
Flask == 2.3.2
google-cloud-datastore == 2.18.0
google-cloud-storage
python-jose
six
requests
authlib